Consolidate more common failure tests in CommonPackingTests
